+++ src/shared/ui/lottieLoading.tsx
